+++ axelor-stock/src/main/java/com/axelor/apps/stock/web/StockMoveLineController.java
@@ -47,9 +47,7 @@ public class StockMoveLineController {
       }
     }
     stockMoveLine = stockMoveLineService.compute(stockMoveLine, stockMove);
-    response.setValue("unitPriceUntaxed", stockMoveLine.getUnitPriceUntaxed());
-    response.setValue("unitPriceTaxed", stockMoveLine.getUnitPriceTaxed());
-    response.setValue("companyUnitPriceUntaxed", stockMoveLine.getCompanyUnitPriceUntaxed());
+    response.setValue("valuatedUnitPrice", stockMoveLine.getValuatedUnitPrice());
   }
 
   public void setProductInfo(ActionRequest request, ActionResponse response) {